Deployed Contracts

+++ 002/contracts/Box.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.8;
 
 contract Box {
-    uint public num;
+    uint256 public num;
 
     function initialize(uint256 _num) external {
         num = _num;
